Zero SpotBugs Issues!

+++ CodenameOne/src/com/codename1/ui/util/UIBuilder.java
@@ -926,7 +926,7 @@ public class UIBuilder { //implements Externalizable {
         if (type == Date.class) {
             boolean b = in.readBoolean();
             if (b) {
-                return new Date(in.readInt());
+                return new Date(in.readLong());
             }
         }
